backend/requirements.txt
apify_client==1.7.1
apify_shared==1.1.2
Cerberus==1.3.5
certifi==2024.7.4
dnspython==2.6.1
Flask==3.0.3
httpcore==1.0.5
httpx==0.27.0
itsdangerous==2.2.0
MarkupSafe==2.1.5
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
typing_extensions==4.12.2
Werkzeug==3.0.3